Derive Macro derive_adhoc_macros::Adhoc

#[derive(Adhoc)]
{
    // Attributes available to this derive:
    #[adhoc]
    #[derive_adhoc]
}

Perform ad-hoc templating driven by a data structure

This macro does two things:

It captures the data structure definition, so that it can be used with calls to derive_adhoc!.

If #[derive_adhoc(MyMacro)] attributes are also specified, they are taken to refer to reuseable templates defined with define_derive_adhoc!. Each such MyMacro is applied to the data structure.

You can specify expansion options for each such template application, by writing #[derive_adhoc(MyMacro[OPTIONS,..])], where [OPTIONS,..] is a comma-separated list of expansion options contained within [ ].

Applying a template (derive-adhoc macro) from another crate

#[derive_adhoc(some_crate::MyMacro)] applies an exported (pub) template defined and exported by some_crate.

You can import a template from another crate, so you can apply it with an unqualified name, with use, but you have to refer to the actual pattern macro name derive_adhoc_template_MyMacro:

use other_crate::derive_adhoc_template_TheirMacro;
#[derive(Adhoc)]
#[derive_Adhoc(TheirMacro)]
struct MyStruct { // ...

Captured data structure definition derive_adhoc_driver_TYPE

The data structure is captured by defining a macro_rules macro called derive_adhoc_driver_TYPE, where TYPE is the name of the type that #[derive(Adhoc)] is applied to.

Like all macro_rules macros, this lives at the top level of your crate. So, sadly, its scoping and importing doesn’t follow that of your actual data type TYPE. If you want to use it outside the containing module, it must be explicitly imported. For complicated reasons to do with Rust’s macro name resolution, a use crate::* won’t necessarily work. You may need to mention it by name explicitly.

There are also ordering requirements: within a crate, the data structure definition may need to precede its uses, in crate and module lexical order. (This can even be affected by the order of modules within a crate.)

#[adhoc] attribute

The contents of #[adhoc] attributes are made available to templates via the ${Xmeta} expansions.

If the template(s) don’t use them, they are ignored. derive-adhoc does not impose any namespacing within #[adhoc]: all templates see the same adhoc meta attributes.

Exporting the driver for downstream crates’ templates

To cause the macro embodying the driver struct to be exported, write: #[derive_adhoc(pub)]. The driver can then be derived from in other crates, with derive_adhoc! { exporting_crate::DriverStruct: ... }.

This is a tricky feature, which should only be used by experts who fully understand the implications. It effectively turns the body of the struct into a macro, with a brittle API and very limited support for namespacing or hygiene.

See pub mod a_driver in the example file pub-a.rs, in the source tree, for a fuller discussion of the implications, and some advice.